Move encode and decode to separate routes to see if we can get around the cloudflare size limits
